requirements.txt
flask
flask-jwt-extended
flask-sqlalchemy
flask_cors
mysql-connector-python
python-dotenv
sqlalchemy
